Move TCP tester code to separate package
