+++ java/src/com/ibm/streamsx/topology/internal/tester/ops/TesterSink.java
@@ -22,8 +22,8 @@ import com.ibm.streams.operator.model.Libraries;
 import com.ibm.streams.operator.model.Parameter;
 import com.ibm.streams.operator.model.PrimitiveOperator;
 import com.ibm.streams.operator.samples.patterns.TupleConsumer;
-import com.ibm.streamsx.topology.internal.tester.TCPTestClient;
-import com.ibm.streamsx.topology.internal.tester.TestTuple;
+import com.ibm.streamsx.topology.internal.tester.tcp.TCPTestClient;
+import com.ibm.streamsx.topology.internal.tester.tcp.TestTuple;
 
 @PrimitiveOperator
 @InputPortSet
