+++ java/src/com/ibm/streamsx/topology/internal/tester/tcp/TCPTestClient.java
@@ -2,7 +2,7 @@
 # Licensed Materials - Property of IBM
 # Copyright IBM Corp. 2015  
  */
-package com.ibm.streamsx.topology.internal.tester;
+package com.ibm.streamsx.topology.internal.tester.tcp;
 
 import java.net.InetSocketAddress;
